Modernize tooling, clean up deps, extract globe into standalone package
- Remove unused deps (postprocessing, d3-geo, gh-pages, redux, CRA leftovers)
- Move @types/* and build tooling to devDependencies
- Migrate from Yarn to Bun (resolutions → overrides, bun.lock, makefile)
- Update .nvmrc to Node 24
- Remove unused Redux store (empty state, no consumers)
- Extract 3D globe into standalone `the-globe` package at ../the-globe
- Decouple EarthScene from trip data (accepts arcs as prop)
- Separate Trip type into tripTypes.ts (app-specific, not globe-specific)
- Create tripArcs.ts for trip-specific arc utilities
- Delete 9 unused landDots JSON variants (~25MB)
- Delete generate-globe/ (moved to the-globe)
- Add AGENTS.md
