+++ .gitignore
@@ -19,8 +19,6 @@
 .env.production.local
 
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
 
 .vscode
 .temp_build/
