+++ .nvmrc
@@ -1,1 +1,1 @@
-20
+24
