--- generate-globe/countries/ne_10m_admin_0_countries.VERSION.txt
@@ -1,1 +1,0 @@
-5.1.1
