--- generate-globe/countries/ne_10m_admin_0_countries.cpg
@@ -1,1 +1,0 @@
-UTF-8
