+++ package.json
@@ -4,52 +4,27 @@
   "private": true,
   "homepage": "https://strelsov.me",
   "dependencies": {
-    "@mdx-js/language-service": "^0.6.1",
     "@mdx-js/react": "^3.1.0",
     "@mdx-js/rollup": "^3.1.0",
-    "@mdx-js/typescript-plugin": "^0.0.7",
     "@nextui-org/react": "^2.6.8",
     "@nextui-org/use-theme": "^2.1.1",
     "@react-spring/three": "^9.7.5",
     "@react-three/drei": "^9.120.4",
     "@react-three/fiber": "^8.17.10",
-    "@react-three/postprocessing": "^2.16.3",
-    "@reduxjs/toolkit": "^2.5.0",
     "@tailwindcss/typography": "^0.5.15",
-    "@turf/turf": "^7.1.0",
-    "@types/jest": "^29.5.14",
-    "@types/mdx": "^2.0.13",
-    "@types/node": "^22.10.2",
-    "@types/prettier": "^3.0.0",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/react-syntax-highlighter": "^15.5.13",
-    "@types/three": "^0.171.0",
-    "@types/uuid": "^10.0.0",
-    "@vitejs/plugin-react": "^4.3.4",
-    "c8": "^10.1.3",
-    "d3-geo": "^3.1.1",
+    "the-globe": "file:../the-globe",
     "framer-motion": "^11.15.0",
-    "gh-pages": "^6.2.0",
     "gsap": "^3.12.5",
-    "jsdom": "^25.0.1",
-    "r3f-perf": "^7.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-flagpack": "^2.0.6",
     "react-parallax": "^3.5.1",
-    "react-redux": "^9.2.0",
     "react-router-dom": "^7.1.0",
     "react-syntax-highlighter": "^15.6.1",
     "react-use-gesture": "^9.1.3",
     "tailwindcss": "^3.4.17",
     "three": "^0.171.0",
-    "ts-jest": "^29.2.5",
-    "vite": "^6.0.6",
-    "vite-plugin-svgr": "^4.3.0",
-    "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^2.1.8",
+    "uuid": "^11.1.0",
     "web-vitals": "^4.2.4"
   },
   "scripts": {
@@ -58,7 +33,7 @@
     "build": "tsc && vite build",
     "serve": "vite preview"
   },
-  "resolutions": {
+  "overrides": {
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "@types/react": "18.2.21",
@@ -78,15 +53,33 @@
   },
   "devDependencies": {
     "@babel/core": "^7.26.0",
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@mdx-js/language-service": "^0.6.1",
+    "@mdx-js/typescript-plugin": "^0.0.7",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
+    "@types/jest": "^29.5.14",
+    "@types/mdx": "^2.0.13",
+    "@types/node": "^22.10.2",
+    "@types/prettier": "^3.0.0",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "18.2.7",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-syntax-highlighter": "^15.5.13",
+    "@types/three": "^0.171.0",
+    "@types/uuid": "^10.0.0",
+    "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
-    "cra-template": "^1.2.0",
+    "c8": "^10.1.3",
+    "jsdom": "^25.0.1",
     "postcss": "^8.4.49",
-    "semver": "^7.6.3",
+    "r3f-perf": "^7.2.3",
+    "ts-jest": "^29.2.5",
     "typescript": "^5.7.2",
-    "typescript-plugin-css-modules": "^5.1.0"
+    "typescript-plugin-css-modules": "^5.1.0",
+    "vite": "^6.0.6",
+    "vite-plugin-svgr": "^4.3.0",
+    "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "^2.1.8"
   }
 }
