--- src/store/store.ts
@@ -1,13 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import globeSlice from "./globeSlice";
-
-const store = configureStore({
-  reducer: {
-    globe: globeSlice,
-  },
-});
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-
-export default store;
